Moved the code of the helper in to the model class + PHP CS corrections
